+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Removed time limit when sending responses
+* Added body HTML message to redirect response
 
 ## v0.1.3 (2023-10-31)
 * Fixed IP extract Request type